requirements.txt
# Core Dependencies
mcp>=1.5.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pyjwt>=2.6.0

# Sanity CMS Dependencies
# Using httpx for Sanity API requests

# Privy Authentication Dependencies
privy-sdk>=0.1.0

# BASE Blockchain Dependencies
web3>=6.0.0
eth-account>=0.8.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
